requirements.txt
aiogram==3.22.0
aiohttp==3.13.2
aiosqlite==0.21.0
beautifulsoup4==4.14.2
lxml==6.0.2
python-dotenv==1.2.1
